// generated/src/app/api-models/cap-pool-stoploss.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapPoolStoploss {
  capModelId: string;
  capEntityCode: string;
  capEntityId: string;
  seqCapPoolId: number;
  capStoplossId: string;
  companyCodeStop: string;
  glRefCodeStop: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  postingRuleStop: string;
  capPoolId: string;
}